fix: map tooltip condition
