+++ src/containers/map/components/map.tsx
@@ -1,26 +1,9 @@
 import { memo } from "react";
-import { useTranslation } from "react-i18next";
-import Map, { type LngLat, Popup } from "react-map-gl";
-
-import { Close } from "@mui/icons-material";
-import {
-  Box,
-  Chip,
-  Divider,
-  IconButton,
-  Paper,
-  Table,
-  TableBody,
-  TableCell,
-  TableRow,
-  Typography,
-} from "@mui/material";
+import Map, { type LngLat } from "react-map-gl";
 
 import { MapView } from "@types";
 
-import { useAppDispatch, useAppSelector } from "@hooks/context";
-
-import { setPopupInfo } from "@context/map";
+import { useAppSelector } from "@hooks/context";
 
 // import { isochrones_selector } from "@context/isochrones/isochrones-selector";
 import { API_TOKEN, MAPBOX_TOKEN } from "@constants";
@@ -30,6 +13,7 @@ import { API_TOKEN, MAPBOX_TOKEN } from "@constants";
 import Isochrones from "./isochrones";
 import Layers from "./layers";
 import LayersDeck from "./layers-deck";
+import PopupTooltip from "./popup";
 
 import "mapbox-gl/dist/mapbox-gl.css";
 
@@ -41,12 +25,9 @@ interface MapProps {
 }
 
 function MapComponent(props: MapProps) {
-  const dispatch = useAppDispatch();
-  const { t } = useTranslation();
-
   const mapStyle = useAppSelector((state) => state.map.style);
-  const popupInfo = useAppSelector((state) => state.map.popupInfo);
   const cursor_mode = props.picking_mode ? "crosshair" : "default";
+
   return (
     <Map
       id="map"
@@ -69,71 +50,8 @@ function MapComponent(props: MapProps) {
       <LayersDeck />
       <Isochrones />
       <Layers />
-      {/** Popup */}
-      {popupInfo && (
-        <Popup
-          style={{
-            zIndex: 20,
-            minWidth: "220px",
-            padding: "0px",
-          }}
-          maxWidth="320px"
-          anchor="bottom"
-          longitude={Number(popupInfo.longitude)}
-          latitude={Number(popupInfo.latitude)}
-          closeButton={false}
-          onClose={() => dispatch(setPopupInfo(null))}
-        >
-          <Paper
-            elevation={0}
-            sx={{
-              maxWidth: 300,
-            }}
-          >
-            <Box position="absolute" top={5} right={5}>
-              <IconButton onClick={() => dispatch(setPopupInfo(null))}>
-                <Close />
-              </IconButton>
-            </Box>
-            <Typography sx={{ m: 1 }} variant="h4">
-              {popupInfo.title}
-            </Typography>
-            <Divider sx={{ mb: 1 }} />
-            <Table size="small">
-              <TableBody>
-                {popupInfo.content &&
-                  Object.keys(popupInfo.content)
-                    .filter((k) => k !== "color")
-                    .map((key) => (
-                      <TableRow key={popupInfo.uid}>
-                        <TableCell align="left" component="th" scope="row">
-                          {t(`popup.${key}`).toUpperCase()}
-                        </TableCell>
-                        {!popupInfo.content["score"] && (
-                          <TableCell align="left">{popupInfo.content[key]}</TableCell>
-                        )}
-                        {popupInfo.content["score"] && (
-                          <Chip
-                            label={popupInfo.content[key]}
-                            clickable={false}
-                            sx={{
-                              mb: 1,
-                            }}
-                            style={{
-                              backgroundColor: `rgb(${popupInfo.content[
-                                "color"
-                              ].toString()})`,
-                            }}
-                          />
-                        )}
-                        <Divider />
-                      </TableRow>
-                    ))}
-              </TableBody>
-            </Table>
-          </Paper>
-        </Popup>
-      )}
+      {/** Popup Tooltip */}
+      <PopupTooltip />
     </Map>
   );
 }
